Update online matching logic

+++ GC-local-server-rewrite/models/OnlineMatchingResultData.cs
@@ -6,51 +6,51 @@ namespace GCLocalServerRewrite.models;
 public class OnlineMatchingResultData
 {
     [ChoXmlElementRecordField(FieldName = "event_id")]
-    public int EventId { get; set; }
+    public long EventId { get; set; }
     
     [ChoXmlElementRecordField(FieldName = "matching_id")]
-    public int MatchingId { get; set; }
+    public long MatchingId { get; set; }
     
     [ChoXmlElementRecordField(FieldName = "class_id")]
-    public int ClassId { get; set; }
+    public long ClassId { get; set; }
     
     [ChoXmlElementRecordField(FieldName = "group_id")]
-    public int GroupId { get; set; }
+    public long GroupId { get; set; }
     
     [ChoXmlElementRecordField(FieldName = "result_score")]
-    public int ResultScore { get; set; }
+    public long ResultScore { get; set; }
     
     [ChoXmlElementRecordField(FieldName = "result_star")]
-    public int ResultStar { get; set; }
+    public long ResultStar { get; set; }
     
     [ChoXmlElementRecordField(FieldName = "result_rank")]
-    public int ResultRank { get; set; }
+    public long ResultRank { get; set; }
     
     [ChoXmlElementRecordField(FieldName = "music_id_1st")]
-    public int MusicIdFirst { get; set; }
+    public long MusicIdFirst { get; set; }
     
     [ChoXmlElementRecordField(FieldName = "music_id_2nd")]
-    public int MusicIdSecond { get; set; }
+    public long MusicIdSecond { get; set; }
     
     [ChoXmlElementRecordField(FieldName = "music_id_3rd")]
-    public int MusicIdThird { get; set; }
+    public long MusicIdThird { get; set; }
     
     [ChoXmlElementRecordField(FieldName = "difficulty_lv_1st")]
-    public int DifficultyFirst { get; set; }
+    public long DifficultyFirst { get; set; }
     
     [ChoXmlElementRecordField(FieldName = "difficulty_lv_2nd")]
-    public int DifficultySecond { get; set; }
+    public long DifficultySecond { get; set; }
     
     [ChoXmlElementRecordField(FieldName = "difficulty_lv_3rd")]
-    public int DifficultyThird { get; set; }
+    public long DifficultyThird { get; set; }
     
     [ChoXmlElementRecordField(FieldName = "item_id_1st")]
-    public int ItemIdFirst { get; set; }
+    public long ItemIdFirst { get; set; }
     
     [ChoXmlElementRecordField(FieldName = "item_id_2nd")]
-    public int ItemIdSecond { get; set; }
+    public long ItemIdSecond { get; set; }
     
     [ChoXmlElementRecordField(FieldName = "item_id_3rd")]
-    public int ItemIdThird { get; set; }
+    public long ItemIdThird { get; set; }
 
 }
